+++ resources/views/admin/subjects/index.blade.php
@@ -26,7 +26,9 @@
                                     <th>ID</th>
                                     <th>Номи Фан</th>
                                     <th>Статус</th>
-                                    <th>Амал</th>
+                                    @if(in_array(auth()->user()->role,['admin','superadmin'] ))
+                                        <th>Амал</th>
+                                    @endif
                                 </tr>
                                 </thead>
                                 <tbody>
@@ -37,14 +39,17 @@
                                         <td>{{$subject->name}}</td>
                                         <td>{!!  $subject->status == 1 ? '<span class="badge badge-success">Active</span>' :
                                                                 '<span class="badge badge-secondary">Inactive</span>'!!}</td>
-                                        <td>
-                                            <a class="btn btn-primary"
-                                               href="{{route('subjects.edit',$subject->id)}}"><i
-                                                    class="fa fa-edit"></i></a>
-                                            <a class="btn btn-danger"
-                                               onclick="deleteSubjectHandler(event,{{$subject->id}})"> <i
-                                                    class="fa fa-trash-o"></i></a>
-                                        </td>
+
+                                        @if(in_array(auth()->user()->role,['admin','superadmin'] ))
+                                            <td>
+                                                <a class="btn btn-primary"
+                                                   href="{{route('subjects.edit',$subject->id)}}"><i
+                                                        class="fa fa-edit"></i></a>
+                                                <a class="btn btn-danger"
+                                                   onclick="deleteSubjectHandler(event,{{$subject->id}})"> <i
+                                                        class="fa fa-trash-o"></i></a>
+                                            </td>
+                                        @endif
                                     </tr>
                                 @endforeach
 
@@ -103,6 +108,7 @@
                     }
                 });
         }
+
         function deleteSubjectHandler(e, id) {
             e.preventDefault();
             swal({
